+++ sys/arch/m68k/include/ansi.h
@@ -50,7 +50,7 @@
 #define	_BSD_PTRDIFF_T_		int		/* ptr1 - ptr2 */
 #define	_BSD_SIZE_T_		unsigned int	/* sizeof() */
 #define	_BSD_SSIZE_T_		int		/* byte count or error */
-#define	_BSD_TIME_T_		long		/* time() */
+#define	_BSD_TIME_T_		long long int	/* time() */
 #if __GNUC_PREREQ__(3,0)
 #define	_BSD_VA_LIST_		__builtin_va_list	/* va_list */
 #else
