+++ sys/dev/pci/pccbb.c
@@ -1371,8 +1371,9 @@ pccbb_power(struct pccbb_softc *sc, int command)
 	splx(s);
 	microtime(&after);
 	timersub(&after, &before, &diff);
-	aprint_debug_dev(sc->sc_dev, "wait took%s %ld.%06lds\n",
-	    (on && times < 0) ? " too long" : "", diff.tv_sec, diff.tv_usec);
+	aprint_debug_dev(sc->sc_dev, "wait took%s %lld.%06lds\n",
+	    (on && times < 0) ? " too long" : "",
+	    (long long int)diff.tv_sec, (long int)diff.tv_usec);
 
 	/*
 	 * Ok, wait a bit longer for things to settle.
